pyproject.toml
[project]
name = "dataset-prepare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "markdown>=3.10.2",
    "numpy>=2.4.4",
    "openpyxl>=3.1.5",
    "pdf2image>=1.17.0",
    "pillow>=12.2.0",
    "playwright>=1.58.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "weasyprint>=68.1",
]
